compilation fix
git-svn-id: https://ffdshow-tryout.svn.sourceforge.net/svnroot/ffdshow-tryout/trunk@1619 3b938f2f-1a1a-0410-8054-a526ea5ff92c

+++ src/xiph/ogg/ogg.h
@@ -21,11 +21,11 @@
 extern "C" {
 #endif
 
-#include <ogg/os_types.h>
+#include <xiph/ogg/os_types.h>
 
 typedef struct {
   void *iov_base; 
-  size_t iov_len;
+  unsigned int iov_len;
 } ogg_iovec_t;
 
 typedef struct {
